DestroySpawnRektEvent
This script is used for the spawning of particles or prefabs in a unity scene.It is used also for detroying an object. For example if we want to make a barell explode we should instantiate fire , smoke and of course destroy the barell object to siulate an explosion. Of course this script can be used only to instantiate particles and prefabs as mentioned in the beggining.