Stream
This script is used for calculating the FPS (frames per second).The code for calculating the FPS is taken from here http://wiki.unity3d.com/index.php?title=FramesPerSecond . When the benchmark starts running a txt file is created at C: where the time and the FPS and time is being saved for further use. This script is responisbe for this action. Moreover it calculates the maximum fps , minimum fps, average fps and number of frames.Lastly when the benchmark is finished a window appears with the cpu ,gpu and operting system of the computer as long as with a score that the pc has achieved.