Benchamrk Unity 3d
An application benchmark created with Unity 3d game engine.This project was created for my graduation project which is required to get my bachelor degree .The benchmark. The benchmark consists of two scenarios one CPU and one GPU benchmark. At the end of each scenario a score is given to the computer system.  This benchmark is functional because we have compared our results with other benchmarks on the other tested on the same computers. The code was written in Uniscript(javascript) and C#